[DSCP-105] Decouple/move Resources, reduce duplication in Params
