+++ src/Dscp/Educator.hs
@@ -1,6 +1,7 @@
 module Dscp.Educator (module M) where
 
 import Dscp.Educator.Block as M
+import Dscp.Educator.CLI as M
 import Dscp.Educator.Launcher as M
 import Dscp.Educator.Serialise ()
 import Dscp.Educator.Txs as M
