+++ src/Dscp/Educator/Launcher/Resource.hs
@@ -1,4 +1,3 @@
-
 -- | Resources used by Educator node
 
 module Dscp.Educator.Launcher.Resource
@@ -9,7 +8,7 @@ import Universum
 
 import Dscp.DB.SQLite (SQLiteDB)
 import Dscp.Educator.Launcher.Params (EducatorParams (..))
-import Dscp.Launcher.Resource (AllocResource (..))
+import Dscp.Resource.Class (AllocResource (..))
 import qualified Dscp.Witness.Launcher.Resource as Witness
 
 -- | Datatype which contains resources required by all Disciplina nodes
