+++ src/Dscp/Educator/Launcher/Runner.hs
@@ -1,4 +1,3 @@
-
 -- | Helpers for starting an Educator node
 
 module Dscp.Educator.Launcher.Runner where
@@ -10,8 +9,8 @@ import Control.Monad.Component (runComponentM)
 import Dscp.Educator.Launcher.Mode (EducatorContext (..), EducatorRealMode)
 import Dscp.Educator.Launcher.Params (EducatorParams (..))
 import Dscp.Educator.Launcher.Resource (EducatorResources (..))
-import Dscp.Launcher.Resource (AllocResource (..))
 import Dscp.Launcher.Rio (runRIO)
+import Dscp.Resource.Class (AllocResource (..))
 import Dscp.Witness.Launcher.Runner (formWitnessContext)
 
 -- | Make up Educator context from dedicated pack of allocated resources.
