+++ src/Dscp/Messages.hs
@@ -81,4 +81,3 @@ serialisePacking = Msg.Packing
 
 -- | Choose 'SerialiseP' as default packing mechanism.
 type Packing = SerialiseP
-
