+++ src/Dscp/Witness.hs
@@ -1,7 +1,7 @@
-
 module Dscp.Witness (module M) where
 
 import Dscp.Accounts.Account as M
+import Dscp.Witness.CLI as M
 import Dscp.Witness.Instances ()
 import Dscp.Witness.Internal as M
 import Dscp.Witness.Launcher as M
