+++ src/educator/Main.hs
@@ -1,4 +1,3 @@
-
 -- | Starting point for running an Educator node
 
 module Main where
@@ -6,25 +5,20 @@ module Main where
 import Universum
 
 import Loot.Log (logInfo, logWarning, modifyLogName)
+import Options.Applicative (execParser, fullDesc, helper, info, progDesc)
 
-import Dscp.DB (RocksDBParams (..), SQLiteDBLocation (..), SQLiteParams (..))
-import Dscp.Educator (EducatorParams (..), launchEducatorRealMode)
-import Dscp.Witness (WitnessParams (..))
-
-import qualified EducatorParams as Params
+import Dscp.CLI (versionOption)
+import Dscp.Educator (EducatorParams (..), educatorParamsParser, launchEducatorRealMode)
 
 main :: IO ()
 main = do
-    Params.EducatorParams {..} <- Params.getEducatorParams
-    let educatorParams = EducatorParams
-            { epWitnessParams = WitnessParams
-                { wpLoggingParams = epLogParams
-                , wpDBParams = RocksDBParams{ rdpPath = epRocksDbPath }
-                }
-            , epDBParams = SQLiteParams
-                { sdpLocation = SQLiteReal epSqliteDbPath }
-            }
+    educatorParams <- getEducatorParams
     launchEducatorRealMode educatorParams $
       modifyLogName (<> "node") $ do
         logInfo "This is the stub for Educator node executable"
         logWarning "Please don't forget to implement everything else!"
+
+getEducatorParams :: IO EducatorParams
+getEducatorParams =
+    execParser $ info (helper <*> versionOption <*> educatorParamsParser) $
+    fullDesc <> progDesc "Discplina educator node."
